% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_table.R
\name{delete_table}
\alias{delete_table}
\title{Delete database table}
\usage{
delete_table(table, database, server)
}
\arguments{
\item{table}{\code{string}. The table to be deleted.}

\item{database}{\code{string}. The database to delete the table from.}

\item{server}{\code{string}. The server holding the database.}
}
\value{
\code{null}
}
\description{
\code{delete_table} opens an ODBC connection and deletes a specified table.
}
\examples{

\dontrun{
delete_table("table name", "database name", "server name")
}

}
