% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_list_tables.R
\name{sdr_list_tables}
\alias{sdr_list_tables}
\title{List tables in a SQL Server database}
\usage{
sdr_list_tables(database, server)
}
\arguments{
\item{database}{\code{string}. A SQL Server database name.}

\item{server}{\code{string}. A SQL Server database server.}
}
\value{
\code{null}
}
\description{
\code{sdr_list_tables} lists all base tables in a SQL Server database on a specified server.
}
\examples{

\dontrun{
sdr_list_tables(database = "DatabaseName", server = "DatabaseServer")
}

}
