% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_read_table.R
\name{sdr_read_table}
\alias{sdr_read_table}
\title{Read a SQL Server Table}
\usage{
sdr_read_table(database, server, table)
}
\arguments{
\item{database}{\code{string}. The database containing the table.}

\item{server}{\code{string}. The server holding the database.}

\item{table}{\code{string}. The table to read data from}
}
\value{
\code{null}
}
\description{
\code{sdr_list_tables} reads the contents of an entire table from a SQL Server database to R.
}
\examples{

\dontrun{
sdr_list_tables(database = "DatabaseName", server = "ServerName", table = "TableName")
}

}
