% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_upload.R
\name{sdr_upload}
\alias{sdr_upload}
\title{Upload Excel worksheets to SDR}
\usage{
sdr_upload(source, archive, database, server)
}
\arguments{
\item{source}{\code{string}. The path of the directory containing Excel files.}

\item{archive}{\code{string}. The path of the directory to send processed files to.}

\item{database}{\code{string}. The database to write the data to.}

\item{server}{\code{string}. The server holding the database.}
}
\value{
\code{null}
}
\description{
\code{sdr_upload} uploads all Excel worksheets in a specified location to a specified database.
}
\examples{

\dontrun{
sdr_upload(source = "C:\\\\example_source", archive = "C:\\\\example_archive", database = "DatabaseName", server = "ServerName")
}

}
