% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload}
\alias{upload}
\title{Upload Excel worksheets to SDR}
\usage{
upload(source, archive, database, server)
}
\arguments{
\item{source}{\code{string}. The path of the directory containing Excel files.}

\item{archive}{\code{string}. The path of the directory to send processed files to.}

\item{database}{\code{string}. The database to write the data to.}

\item{server}{\code{string}. The server holding the database.}
}
\value{
\code{null}
}
\description{
\code{upload} uploads all Excel worksheets in a specified location to a specified database.
}
\examples{

\dontrun{
upload("C://example_source", "C://example_archive", "database_name", "server_name")
}

}
